final commit.
my buggy, bugs is finally bugged to an end.
and it was really fun, i miss this project, kind of don't want to let it go, but it's time to move on. and likely i will learn most of things i couldn't implement in this project.

it is notes and tasks app, with a MainFragment.kt that hold references to both list in a recycler view, and it's also where the problems began, as i couldn't merge two livedatas into one and submit only one list to the differ.

a lot can be said, but am tired
during this 29 days, time that took me to come up with what is so far, my biggest project from scratch, i understood the difference of working on a complete app and just building something,

when one is just building an app, it is rather "fast", the goal is to get the app up and running and that's all, but when you dedicate yourself into building features and making the app what you think an of it's kind should be able to do and feel, then you understand why "it can take a week to change the style of button", I say this because in big companies no one is building the app from scratch, rather they are either implementing new features into a massive bunch of code or they are updating features in a massive bunch of code.

i would like to say, that i won't go so deep into my next project (messaging app) but now i believe that, that's where the harder learning curve is.